ADEN-11129 move ad label/loading logic to shared/utils
