+++ platforms/ucp-mobile/setup/dynamic-slots/ucp-mobile-dynamic-slots.setup.ts
@@ -1,4 +1,4 @@
-import { slotsContext } from '@platforms/shared';
+import { removeAdLabel, slotsContext } from '@platforms/shared';
 import {
 	AdSlot,
 	adSlotEvent,
@@ -24,21 +24,6 @@ import {
 	SlotSetupDefinition,
 	UcpMobileSlotsDefinitionRepository,
 } from './ucp-mobile-slots-definition-repository';
-
-const removeLabel = (slotName: string): void => {
-	const parentElement =
-		slotName !== 'top_leaderboard'
-			? document.querySelector(`#${slotName}`).parentElement
-			: document.querySelector('.top-ads-container');
-
-	let adLabel: HTMLElement;
-	for (const child of parentElement.children as any) {
-		if (child.className.includes('ae-translatable-label')) {
-			adLabel = child;
-		}
-	}
-	adLabel?.classList.add('hide');
-};
 
 @Injectable()
 export class UcpMobileDynamicSlotsSetup implements DiProcess {
@@ -137,7 +122,7 @@ export class UcpMobileDynamicSlotsSetup implements DiProcess {
 				.subscribe((action) => {
 					removeLoader(action.adSlotName);
 					if (action['event'] === 'slotHidden') {
-						removeLabel(action.adSlotName);
+						removeAdLabel(action.adSlotName);
 					}
 				});
 		};
@@ -177,7 +162,7 @@ export class UcpMobileDynamicSlotsSetup implements DiProcess {
 				.subscribe((action) => {
 					removeLoader(action.adSlotName);
 					if (action['event'] === 'slotHidden') {
-						removeLabel(action.adSlotName);
+						removeAdLabel(action.adSlotName);
 					}
 				});
 		};
